+++ lib/provider/hive/application_settings.dart
@@ -34,6 +34,7 @@ class ApplicationSettingsHiveProvider
   void registerAdapters() {
     Log.debug('registerAdapters()', '$runtimeType');
     Hive.maybeRegisterAdapter(ApplicationSettingsAdapter());
+    Hive.maybeRegisterAdapter(CallButtonsPositionAdapter());
   }
 
   /// Returns the stored [ApplicationSettings] from [Hive].
@@ -147,6 +148,17 @@ class ApplicationSettingsHiveProvider
     await putSafe(
       0,
       (box.get(0) ?? ApplicationSettings())..pinnedActions = buttons,
+    );
+  }
+
+  /// Stores a new [position] value of
+  /// [ApplicationSettings.callButtonsPosition] to [Hive].
+  Future<void> setCallButtonsPosition(CallButtonsPosition position) async {
+    Log.debug('setCallButtonsPosition($position)', '$runtimeType');
+
+    await putSafe(
+      0,
+      (box.get(0) ?? ApplicationSettings())..callButtonsPosition = position,
     );
   }
 }