+++ CHANGELOG.md
@@ -11,6 +11,12 @@ All user visible changes to this project will be documented in this file. This p
 
 [Diff](/../../compare/v0.1.0-alpha.11.1...v0.1.0-alpha.12) | [Milestone](/../../milestone/14)
 
+### Changed
+
+- UI:
+    - Chat page:
+        - Call buttons position setting and adjusting. ([#750], [#718])
+
 ### Fixed
 
 - UI:
@@ -19,6 +25,8 @@ All user visible changes to this project will be documented in this file. This p
     - Contacts tab:
         - Restore button displaying under mobile navigation bar. ([#763, #758])
 
+[#718]: /../../issues/718
+[#750]: /../../pull/750
 [#758]: /../../issues/758
 [#763]: /../../pull/763